+++ .gitignore
@@ -57,3 +57,6 @@ buck-out/
 
 # CocoaPods
 /ios/Pods/
+
+ios/NUSSGH_Empower.xcworkspace/xcshareddata/WorkspaceSettings.xcsettings
+
